+++ src/pyjion/absint.h
@@ -349,7 +349,7 @@ private:
     void ensureLabels(vector<Label>& labels, size_t count);
 
     void branchRaise(const char* reason = nullptr, size_t curByte = ~0);
-    void raiseOnNegativeOne(int curByte);
+    void raiseOnNegativeOne(size_t curByte);
 
     void unwindEh(ExceptionHandler* fromHandler, ExceptionHandler* toHandler = nullptr);
